changed data type from Long to Integer of the id in InvoiceModel

+++ src/main/java/com/example/ContaGest/controller/ClientController.java
@@ -4,9 +4,9 @@ import com.example.ContaGest.dto.SaveImage;
 import com.example.ContaGest.service.ClientService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/client")
@@ -17,11 +17,11 @@ public class ClientController {
 
     @PostMapping("/save-img")
     public void saveImage(SaveImage saveImage) throws IOException {
-        clientService.saveImage(saveImage.getClientCI(),saveImage.getMonth(),saveImage.getFile().getBytes());
+        clientService.saveImage(saveImage.getClientCI(), saveImage.getMonth(), saveImage.getFile().getBytes());
     }
 
     @DeleteMapping("/delete-img/{imgID}")
-    public void deleteImage (@PathVariable Long imgID){
+    public void deleteImage(@PathVariable Integer imgID) {
         clientService.deleteImage(imgID);
     }
